Grep for our merge script to determine branch
Before we were grepping for 'codeaurora' and that was working really well
but since we're not tracking the CAF vendor telephony repo on the AOSP branch,
some users have reported seeing the wrong lunch menu. The bug was confirmed.

This changes what we look for to the merge script. Since we linkfile these scripts
to their respective branch, there shouldn't be a problem. All is well now :-)

Change-Id: Ib2cf30cfc4c411b17424f526f8d98c2f9ca9e8a4